backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# File Processing & AI
PyPDF2==3.0.1
pdf2image==1.17.0
Pillow==10.1.0
easyocr==1.7.0
pytesseract==0.3.10

# Text-to-Speech & Translation
gtts==2.5.0
googletrans==4.0.0rc1
azure-cognitiveservices-speech==1.34.0
boto3==1.34.0  # For AWS Polly

# OpenAI & ML
openai==1.3.5
transformers==4.36.0
torch==2.1.1
huggingface-hub==0.19.4

# HTTP Requests & API
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Email & Notifications
fastapi-mail==1.4.1
twilio==8.11.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# CORS & Middleware
fastapi-cors==0.0.6

# File Storage
minio==7.2.0

# Utils
python-slugify==8.0.1
pytz==2023.3
validators==0.22.0
